samples project can wait, ts can wait. discrete builds
