+++ package.json
@@ -29,5 +29,5 @@
       }
     }
   },
-  "packageManager": "pnpm@7.0.0"
+  "packageManager": "pnpm@7.17.0"
 }
