--- samples/.storybook/preview-head.html
@@ -1,3 +1,0 @@
-<script>
-  window.global = window;
-</script>
