--- samples/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "samples",
-  "private": true,
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
-  },
-  "dependencies": {
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.17.10",
-    "@storybook/addon-actions": "^6.5.0-beta.5",
-    "@storybook/addon-essentials": "^6.5.0-beta.5",
-    "@storybook/addon-interactions": "^6.5.0-beta.5",
-    "@storybook/addon-links": "^6.5.0-beta.5",
-    "@storybook/builder-vite": "^0.1.33",
-    "@storybook/react": "^6.5.0-beta.5",
-    "@storybook/testing-library": "^0.0.11",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
-    "@vitejs/plugin-react": "^1.3.0",
-    "babel-loader": "^8.2.5"
-  }
-}
