--- samples/src/App.tsx
@@ -1,1 +1,0 @@
-export const App = () => <div>Hello World</div>;
