--- samples/src/main.tsx
@@ -1,9 +1,0 @@
-import { StrictMode } from "react";
-import ReactDOM from "react-dom/client";
-import { App } from "./App";
-
-ReactDOM.createRoot(document.getElementById("root")!).render(
-  <StrictMode>
-    <App />
-  </StrictMode>
-);
