--- samples/src/stories/Introduction.stories.mdx
@@ -1,7 +1,0 @@
-import { Meta } from "@storybook/addon-docs";
-
-<Meta title="Intro" />
-
-# Intro
-
-Add stories here
